% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{clean_joined_stats}
\alias{clean_joined_stats}
\title{Clean Joined LPGA Stats}
\usage{
clean_joined_stats(df = NULL)
}
\arguments{
\item{df}{Dataframe.  Created using the stats_join() function.}
}
\description{
This function cleans the dataframe created with the stats_join() function.
}
