% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{clean_ladygolf}
\alias{clean_ladygolf}
\title{Clean LPGA Statistics Dataframes}
\usage{
clean_ladygolf()
}
\description{
Cleans up the dataframes created from lpga_scrape() function.  Returns a list of dataframes.
Input dataframes must have specific names and variable names, created with the stat_names() function
in ladygolfstats.
}
