% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{lpga_getlinks}
\alias{lpga_getlinks}
\title{Get Prior Year LPGA Stats Links}
\usage{
lpga_getlinks(year = 2016)
}
\arguments{
\item{year}{Numeric.  This is the previous year that you want to collect the statistics for.}
}
\description{
This function gets the links for pages that have LPGA statistics for previous years.
Currently only gets previous year pages for statistics that are on the main page
of the statistics section of LPGA.com.
}
