% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{lpga_scrape}
\alias{lpga_scrape}
\title{Scrape LPGA Stats Pages}
\usage{
lpga_scrape(links = NULL)
}
\arguments{
\item{links}{A list of website URLs}
}
\description{
This function takes a list of website pages from LPGA.com for current year or prior years statistics
and pulls the data table. The list can be the output from the lpga_getlinks() function.
 A data frame is created from the table from each page  and added to a list of data frames.
 Output is a list of dataframes.
}
