% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{rolex_clean}
\alias{rolex_clean}
\title{Clean Rolex PoY Dataframe}
\usage{
rolex_clean(year = 2016)
}
\arguments{
\item{year}{The year of the statistic pulled from the LPGA.com site.}
}
\description{
Select the correct table from the Rolex Player of the Year statistics site
on LPGA.com.
}
