% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{stat_names}
\alias{stat_names}
\title{Get LPGA Statistics Names}
\usage{
stat_names(links = NULL, year = NULL)
}
\arguments{
\item{links}{A list of website URLs}
}
\description{
This function is used by other functions in the package.  It extracts the name of the
statistic from the url for all the links gathered using the lpga_getlinks() function.
Could also be used for any list of statistics URLs from the LPGA.com site.
}
