% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_previous_years.R
\name{stats_join}
\alias{stats_join}
\title{Create By Name Statistics Dataframe}
\usage{
stats_join()
}
\description{
This function joins the dataframes created from the lpga_scrape() function into
one dataframe that provides statistics for the year by name of player.  It requires
that the dataframes be in the Environment and named based on the output of the stat_names()
function.
}
